[exynos9810] drivers: modem: Fix different enumeration types
drivers/misc/modem_v1/link_device_memory_flow_control.c:123:14: error: comparison of different enumeration types ('enum legacy_ipc_map' and 'enum dev_format') [-Werror,-Wenum-compare]
  123 |         if (dev->id == IPC_RAW && atomic_read(&dev->txq.busy) == 0) {
      |             ~~~~~~~ ^  ~~~~~~~
drivers/misc/modem_v1/link_device_memory_flow_control.c:165:14: error: comparison of different enumeration types ('enum legacy_ipc_map' and 'enum dev_format') [-Werror,-Wenum-compare]
  165 |         if (dev->id == IPC_RAW && atomic_read(&dev->txq.busy) > 0) {
      |             ~~~~~~~ ^  ~~~~~~~
drivers/misc/modem_v1/link_device_memory_debug.c:121:9: error: comparison of different enumeration types ('enum legacy_ipc_map' and 'enum dev_format') [-Werror,-Wenum-compare]
  121 |         if (id > IPC_RAW)
      |             ~~ ^ ~~~~~~~

error: comparison of different enumeration types ('enum legacy_ipc_map' and 'enum dev_format')

+++ drivers/misc/modem_v1/link_device_memory_debug.c
@@ -118,7 +118,7 @@ void print_dev_snapshot(struct mem_link_device *mld, struct mem_snapshot *mst,
 	struct link_device *ld = &mld->link_dev;
 	enum legacy_ipc_map id = dev->id;
 
-	if (id > IPC_RAW)
+	if ((int)id > IPC_RAW)
 		return;
 
 	mif_err("%s: [%s] %s | TXQ{in:%u out:%u} RXQ{in:%u out:%u} | "
